requirements.txt
# Interface Gráfica
PyQt6==6.7.0
PyQt6-Qt6==6.7.0
PyQt6-sip==13.6.0

# Banco de Dados
supabase==2.10.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1

# Áudio
pyaudio==0.2.14
wave==0.0.2
sounddevice==0.5.1
numpy==1.26.0

# IA e Processamento
openai==1.51.0
tiktoken==0.8.0
langchain==0.3.7
langchain-openai==0.2.3

# Utilitários
python-dateutil==2.9.0
requests==2.32.3
aiohttp==3.10.10
asyncio==3.4.3
loguru==0.7.2

# Desenvolvimento (opcional)
pytest==8.3.3
pytest-asyncio==0.24.0
black==24.10.0
flake8==7.1.1